Fixed UI bug, music pauses when game is paused now
- Fixed issue regarding UI not properly stopping the game and not being interactable.
- Music pauses when the player opens the escape menu or if the player wins/loses.
