+++ CSC311 Project Perry Winch/Assets/Animations/Animation Controllers/BattleDroidAC.controller
@@ -1521,9 +1521,9 @@ AnimatorStateMachine:
   m_EntryTransitions: []
   m_StateMachineTransitions: {}
   m_StateMachineBehaviours: []
-  m_AnyStatePosition: {x: -80, y: -240, z: 0}
+  m_AnyStatePosition: {x: -50, y: -150, z: 0}
   m_EntryPosition: {x: -50, y: -50, z: 0}
-  m_ExitPosition: {x: 110, y: -240, z: 0}
+  m_ExitPosition: {x: -50, y: -230, z: 0}
   m_ParentStateMachinePosition: {x: 800, y: 20, z: 0}
   m_DefaultState: {fileID: -206486415234852734}
 --- !u!1101 &1225095761012251401
