--- CSC311 Project Perry Winch/Assets/Prefabs/Barrier Variant.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d10a565e62bbace47b9e37bcfe8db400
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
