+++ CSC311 Project Perry Winch/Assets/Prefabs/Brazier (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0e6e798034013aa469b984092856e546
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
