+++ CSC311 Project Perry Winch/Assets/Scripts/BrazierFire.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BrazierFire : MonoBehaviour
+{
+    public GameObject fireOne;
+    public GameObject fireTwo;
+
+    private void OnTriggerEnter(Collider other)
+    {
+        if(other.gameObject.name == "Player")
+        {
+            fireOne.SetActive(true);
+            fireTwo.SetActive(true);
+        }
+    }
+}
