+++ CSC311 Project Perry Winch/Assets/Scripts/BrazierFire.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 729fff6dbaf340740a00ef7e9c7ec9f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
